+++ doc/release-notes.md
@@ -15,3 +15,8 @@ in BCH per kilobyte, not satoshi per byte.
 RPC changes
 -----------
 The `getblockstats` RPC is faster for fee calculation by using BlockUndo data. Also, `-txindex` is no longer required and `getblockstats` works for all non-pruned blocks.
+
+Miscellaneous RPC changes
+------------
+
+- `createwallet` can now create encrypted wallets if a non-empty passphrase is specified.